requirements.txt
carvekit_colab==4.1.0
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
requests==2.28.2
flask==1.1.4
markupsafe==2.0.1
matplotlib==3.7.3